+++ CHANGELOG.md
@@ -49,8 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - InputFilter component
-- FilterEditor component
-- InputFilterChip component
+- InputFiltersChipEditor component
+- InputFiltersChip component
 - ActionListFilter component
 
 - `Drawer` component